dealer_sync_backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.25
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
kombu==5.3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.7
requests==2.32.3
selenium==3.141.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
